refactor: RunTest 내용  GameCenterControllerTest 로 이동
